plant-seeding-classification/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymysql==1.1.1
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.26.4